Function to calculate daily values from cumulative values

Calculate daily COVID case counts from cumulative counts

calculate_daily_counts(cum_counts_df)

Arguments

cum_counts_df
A dataframe of cumulative confirmed cases, deaths and recoveries in Australia. Can be downloaded using download_data(data_type = "cumulative")

Value

A dataframe of daily confirmed COVID-19 cases, deaths and recoveries in Australia

Examples

calculate_daily_counts(cum_counts_df = download_data(data_type = "cumulative"))

#> Start downloading JHU CSSE Covid-19 data

#> Downloading cached version of JHU CSSE Covid 19 data...

#> Warning: Unknown or uninitialised column: `timestamp`.

#> 

#> 
#> Data Info:

#> The COVID-19 Data Repository by the Johns Hopkins University Center for

#> Systems Science and Engineering (JHU CSSE) relies upon publicly

#> available data from multiple sources that do not always agree. It is

#> updated daily. The data comes in three data frames that you can select

#> by the 'type' parameter. The 'country' data frame contains the global

#> country-level data reported by JHU CSSE by aggregating over the

#> regional data for countries that have regional data available. The

#> 'country_region' data frame provides regional data for the countries

#> that have regional data available (mosty Australia, Canada and China).

#> The 'us_county' data frame reports the data for the U.S. at the county

#> level. The column 'timestamp' reports the time the data was downloaded

#> from its authoritative source.

#> 

#> For further information refer to:

#> https://github.com/CSSEGISandData/COVID-19.

#> 

#> # A tibble: 2,176 x 6
#> # Groups:   region [8]
#>    iso3c region                       Date       Confirmed Deaths Recovered
#>    <chr> <chr>                        <date>         <dbl>  <dbl>     <dbl>
#>  1 AUS   Australian Capital Territory 2020-01-22         0      0         0
#>  2 AUS   New South Wales              2020-01-22         0      0         0
#>  3 AUS   Northern Territory           2020-01-22         0      0         0
#>  4 AUS   Queensland                   2020-01-22         0      0         0
#>  5 AUS   South Australia              2020-01-22         0      0         0
#>  6 AUS   Tasmania                     2020-01-22         0      0         0
#>  7 AUS   Victoria                     2020-01-22         0      0         0
#>  8 AUS   Western Australia            2020-01-22         0      0         0
#>  9 AUS   Australian Capital Territory 2020-01-23         0      0         0
#> 10 AUS   New South Wales              2020-01-23         0      0         0
#> # … with 2,166 more rows