Generate value box after summing a vector

Generate value box after summing a vector

generate_value_box(counts_vector, text, icon_name, color)

Arguments

counts_vector
A numeric vector, To be summed up in generating the value printed

text
A String, The text printed on the value box

icon_name
A String, The font-awesome icon name

color
A String, The color of the box

Value

An html widget including the summed vector, text and icon

Examples

generate_value_box(counts_vector=c(1, 5, 4, 9, 0),
  text="Confirmed Cases", icon="head-side-mask",
  color="green")

#> <div class="col-sm-4">
#>   <div class="small-box bg-green">
#>     <div class="inner">
#>       <h3>19</h3>
#>       <p>Confirmed Cases</p>
#>     </div>
#>     <div class="icon-large">
#>       <i class="fa fa-head-side-mask"></i>
#>     </div>
#>   </div>
#> </div>

x <- c(37,42,75,29,46,85)
generate_value_box(counts_vector=x, text="Confirmed Cases",
  icon="head-side-mask", color="green")

#> <div class="col-sm-4">
#>   <div class="small-box bg-green">
#>     <div class="inner">
#>       <h3>314</h3>
#>       <p>Confirmed Cases</p>
#>     </div>
#>     <div class="icon-large">
#>       <i class="fa fa-head-side-mask"></i>
#>     </div>
#>   </div>
#> </div>

generate_value_box(get_case_count(data_frame =
  download_data(data_type = "daily"), date = "2020-10-01",
  case_type = "Confirmed"), text="Confirmed Cases",
  icon="head-side-mask", color="green")

#> Start downloading JHU CSSE Covid-19 data

#> Downloading cached version of JHU CSSE Covid 19 data...

#> Warning: Unknown or uninitialised column: `timestamp`.

#> 

#> 
#> Data Info:

#> The COVID-19 Data Repository by the Johns Hopkins University Center for

#> Systems Science and Engineering (JHU CSSE) relies upon publicly

#> available data from multiple sources that do not always agree. It is

#> updated daily. The data comes in three data frames that you can select

#> by the 'type' parameter. The 'country' data frame contains the global

#> country-level data reported by JHU CSSE by aggregating over the

#> regional data for countries that have regional data available. The

#> 'country_region' data frame provides regional data for the countries

#> that have regional data available (mosty Australia, Canada and China).

#> The 'us_county' data frame reports the data for the U.S. at the county

#> level. The column 'timestamp' reports the time the data was downloaded

#> from its authoritative source.

#> 

#> For further information refer to:

#> https://github.com/CSSEGISandData/COVID-19.

#> 

#> <div class="col-sm-4">
#>   <div class="small-box bg-green">
#>     <div class="inner">
#>       <h3>13</h3>
#>       <p>Confirmed Cases</p>
#>     </div>
#>     <div class="icon-large">
#>       <i class="fa fa-head-side-mask"></i>
#>     </div>
#>   </div>
#> </div>